% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download365_file}
\alias{download365_file}
\title{Download files stored in OneDrives or SharePoints}
\usage{
download365_file(where, SharePoint, drive, file)
}
\arguments{
\item{where}{"OneDrive" or "SharePoint"}

\item{SharePoint}{url}

\item{drive}{SharePoint drive}

\item{file}{File path}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function downloads files stored in OneDrives or SharePoints
}
