% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharePoint.R
\name{get_drives}
\alias{get_drives}
\title{Get the drive names of a Sharepoint page}
\usage{
get_drives(SharePoint)
}
\arguments{
\item{SharePoint}{\code{\link[Microsoft365R:client]{Microsoft365R::get_sharepoint_site()}}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function gets the drive names of a Sharepoint page
}
