% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharePoint.R
\name{get_index}
\alias{get_index}
\title{Get the index of a Sharepoint page drive}
\usage{
get_index(SharePoint, drive)
}
\arguments{
\item{SharePoint}{\code{\link[Microsoft365R:client]{Microsoft365R::get_sharepoint_site()}}}

\item{drive}{SharePoint drive}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function gets the index of a Sharepoint page drive
}
