% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneDrive.R
\name{OneDrive_csv}
\alias{OneDrive_csv}
\title{Read '.csv' files stored in an OneDrive}
\usage{
OneDrive_csv(file, ...)
}
\arguments{
\item{file}{'.csv' file path}

\item{...}{\code{\link[=read.csv]{read.csv()}} arguments}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function reads '.csv' files stored in an OneDrive
}
\examples{
\dontrun{OneDrive_csv("path/file.csv")}
}
