% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneDrive.R, R/SharePoint.R
\name{OneDrive_shared_file}
\alias{OneDrive_shared_file}
\alias{SharePoint_shared_file}
\title{Download and read a file from a shared url}
\usage{
OneDrive_shared_file(url, .function, ...)

SharePoint_shared_file(url, .function, ...)
}
\arguments{
\item{url}{A shared url from OneDrive or SharePoint.}

\item{.function}{The function will read the file.}

\item{...}{.function args.}
}
\description{
This function downloads and reads a file from a shared url.
}
\examples{
 url = 'https://arturhgq-my.sharepoint.com/:x:/p/contact/EZ_KJ3cqtIVEh4PdXhTGy7IBpWT5_Zlp2VjYwlgVCPK4oQ?e=UZGQZt'

 OneDrive_shared_file(
    url,
    .function = readxl::read_excel
  )
}
\references{
Based on https://github.com/PaulMelloy/Download_from_OneDrive
}
