% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneDrive.R
\name{OneDrive_xlsx}
\alias{OneDrive_xlsx}
\title{Read '.xlsx' files stored in an OneDrive}
\usage{
OneDrive_xlsx(file, ...)
}
\arguments{
\item{file}{'.xlsx' file path}

\item{...}{\code{\link[readxl:read_excel]{readxl::read_excel()}} arguments}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function reads '.xlsx' files stored in an OneDrive
}
\note{
You may read several 'xlsx' sheets using \code{sheet = c('sheet1', 'sheet2')}
}
\examples{
\dontrun{OneDrive_xlsx("path/file.xlsx")}
}
