% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read365-package.R
\docType{package}
\name{read365-package}
\alias{read365}
\alias{read365-package}
\title{read365: Read OneDrive and SharePoint files}
\description{
This package provides tools to read OneDrive and SharePoint files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/arturhgq/read365}
  \item \url{https://arturhgq.github.io/read365/}
  \item Report bugs at \url{https://github.com/arturhgq/read365/issues}
}

}
\author{
\strong{Maintainer}: Artur Quirino \email{contact@arturhgq.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
