% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharePoint.R
\name{SharePoint_csv}
\alias{SharePoint_csv}
\title{Read '.csv' files stored in a SharePoint}
\usage{
SharePoint_csv(SharePoint, drive, file, ...)
}
\arguments{
\item{SharePoint}{url}

\item{drive}{SharePoint drive}

\item{file}{'.csv' file path}

\item{...}{\code{\link[=read.csv]{read.csv()}} arguments}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function reads '.csv' files stored in a SharePoint
}
\examples{
\dontrun{
SharePoint_csv(
  SharePoint = "https://page.sharepoint.com",
  drive = "Documents",
  file = "Book.csv"
  )
}
}
