% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharePoint.R
\name{SharePoint_file}
\alias{SharePoint_file}
\title{Read files stored in a SharePoint}
\usage{
SharePoint_file(SharePoint, drive, file, .function, ...)
}
\arguments{
\item{SharePoint}{url}

\item{drive}{SharePoint drive}

\item{file}{File path}

\item{.function}{Function that will be used to reading the file}

\item{...}{\code{.function} args}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function reads files stored in a SharePoint
}
\examples{
\dontrun{
SharePoint_file(
  SharePoint = "https://page.sharepoint.com",
  drive = "Documents",
  file = "Book.txt",
  read.delim,
  sep = ""
  )
}
}
